Progress on circular edge calibration -TODO: radial average calculation

+++ fpmcode/WallerCode/calDF.py
@@ -36,5 +36,5 @@ def calDF(FI, XYmid):
 
 
     
-    return DF
+    return DF.astype(int)
